# api/src/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.0.1",
        "title": "Zoo API",
        "description": "API du projet zoo",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": {
        "BearerAuth": []
    },
    "servers": [{ "url": "/api/" }],
...
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/animals/": {
            "get": {
                "tags": ["Animals"],
                "summary": "Lister les infos de tous les animaux du zoo",
                "responses": {
                    "200": {
                        "description": "Succès.",
                        "schema": {
                            "$ref": "#/definitions/Animal"
                        }
                    },
                    "400": {
                        "description": "Syntaxe de la requête érronée."
                    },
                    "404": {
                        "description": "Animal inconnu."
                    }
                }
            },
            "post": {
                "tags": ["Animals"],
                "summary": "Entrer un nouvel animal dans la base",
                "requestBody": {
                    "description": "Objet Animal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Animal"
                            }
                        }
                    }
                },
                "produces": ["application/json"],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Animal entré dans la base.",
                        "schema": {
                            "$ref": "#/definitions/Animal"
                        }
                    },
                    "400": {
                        "description": "Syntaxe de la requête érronée."
                    }
                }
            }
        },
...
        "/employees/": {
            "get": {
                "tags": ["Employees"],
                "summary": "Lister les employés",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès.",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    },
                    "400": {
                        "description": "Syntaxe de la requête érronée."
                    },
                    "404": {
                        "description": "Employé inconnu."
                    }
                }
            }
        },
...
        "/enclosures/": {
            "get": {
                "tags": ["Enclosures"],
                "summary": "Lister les enclos",
                "responses": {
                    "200": {
                        "description": "Succès.",
                        "schema": {
                            "$ref": "#/definitions/Enclosure"
                        }
                    },
                    "400": {
                        "description": "Syntaxe de la requête érronée."
                    },
                    "404": {
                        "description": "Enclos inconnus."
                    }
                }
            }
        },
...
        "/species/": {
            "get": {
                "tags": ["Species"],
                "summary": "Lister toutes les espèces du zoo",
                "responses": {
                    "200": {
                        "description": "Succès.",
                        "schema": {
                            "$ref": "#/definitions/Species"
                        }
                    },
                    "400": {
                        "description": "Syntaxe de la requête érronée."
                    },
                    "404": {
                        "description": "Espèce inconnue."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Animal": {
            "type": "object",
            "properties": {
                "nom": {
                    "type": "string"
                },
                "espece": {
                    "type": "string"
                },
                "naissance": {
                    "type": "string"
                },
                "deces": {
                    "type": "string"
                },
                "sexe": {
                    "type": "string"
                },
                "observations": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "Employee": {
            "type": "object",
            "properties": {
                "nom": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "secu": {
                    "type": "string"
                },
                "naissance": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Enclosure": {
            "type": "object",
            "properties": {
                "nom": {
                    "type": "string"
                },
                "nomApp": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                },
                "coordonnées": {
                    "type": "string"
                },
                "superficie": {
                    "type": "integer"
                }
            }
        },
...
        "Species": {
            "type": "object",
            "properties": {
                "nom": {
                    "type": "string"
                },
                "nomApp": {
                    "type": "string"
                },
                "sociable": {
                    "type": "boolean"
                },
                "observations": {
                    "type": "string"
                },
                "dangereux": {
                    "type": "boolean"
                },
                "enclos": {
                    "type": "string"
                }
            }
        }
    }
}